Change loop counter from int to long
This is a partial solution to raise the limit on cycle count. There are still issues reading in a long type from the serial connection.

+++ Stepper/Stepper.cpp
@@ -98,8 +98,8 @@ void Stepper::move(const double distance) {
 }
 
 // Power the motor with a square wave
-void Stepper::square(const double distance, const int hold, const int cycles) {
-    for(int i = 0; i < cycles; ++i) {
+void Stepper::square(const double distance, const int hold, const long cycles) {
+    for(long i = 0; i < cycles; ++i) {
         move(distance);
         delay(hold);
         move(-distance);
